+++ core/src/com/mygdx/game/Characters/Mudkip.java
@@ -12,7 +12,7 @@ public class Mudkip extends Character{
     PokemonType type;
     
 
-    public Mudkip(int x, int y){
+    public Mudkip(float x, float y){
         this.setPosition( x, y);
         this.setSize(64, 64);
         this.speed = 2;
