+++ core/src/com/mygdx/game/Characters/Pikachu.java
@@ -13,7 +13,7 @@ public class Pikachu extends Character
     PokemonType type;
     
 
-    public Pikachu(int x, int y){
+    public Pikachu(float x, float y){
         this.setPosition( x, y);
         this.setSize(64, 64);
         this.speed = 2;
